--- a/.rspec
+++ b/.rspec
@@ -1,3 +1,4 @@
 --format Fuubar
 --colour
 --color
+--require spec_helper