--- a/spec/controllers/api/v1/activities_controller_spec.rb
+++ b/spec/controllers/api/v1/activities_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Api::V1::ActivitiesController, :type => :controller do