--- a/spec/models/ability_spec.rb
+++ b/spec/models/ability_spec.rb
@@ -1,3 +1,4 @@
+require 'rails_helper'
 require "cancan/matchers"
 
 describe "User", :type => :model do