--- a/spec/models/activity_clasification_spec.rb
+++ b/spec/models/activity_clasification_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ActivityClassification, :type => :model do