--- a/spec/models/activity_session_spec.rb
+++ b/spec/models/activity_session_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 
 describe ActivitySession, :type => :model do