--- a/spec/models/activity_spec.rb
+++ b/spec/models/activity_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Activity, :type => :model do