--- a/spec/models/assignment_view_spec.rb
+++ b/spec/models/assignment_view_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe AssignmentView, :type => :model do