--- a/spec/models/chapter_shim_spec.rb
+++ b/spec/models/chapter_shim_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ChapterShim, :type => :model do