--- a/spec/models/classroom_spec.rb
+++ b/spec/models/classroom_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Classroom, :type => :model do