--- a/spec/models/district_spec.rb
+++ b/spec/models/district_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe District, type: :model do
   describe 'setup from clever', :vcr do