--- a/spec/models/district_spec.rb
+++ b/spec/models/district_spec.rb
@@ -1,39 +1,43 @@
 require 'rails_helper'
 
 describe District, type: :model do
+  def other_object_counts
+    "# Classrooms: #{Classroom.count}; # Users: #{User.count}"
+  end
+
   describe 'setup from clever', :vcr do
     it 'passes an auth hash for a district to be setup' do
-      @district = District.setup_from_clever({
-        info: {
-          name: 'Fake District',
-          id: '535e91b05fc8cb160e001645'
-        },
-        credentials: {
-          token: '123'
-        }
-      })
+      def setup_district
+        @district = District.setup_from_clever({
+          info: {
+            name: 'Fake District',
+            id: '535e91b05fc8cb160e001645'
+          },
+          credentials: {
+            token: '123'
+          }
+        })
+      end
+
+      expect { setup_district }.not_to change { other_object_counts }
 
       expect(@district.valid?).to be_truthy
       expect(@district.id).to_not be_nil
       expect(@district.name).to eq('Fake District')
-
-      expect(Classroom.count).to eq(4)
-      expect(User.teacher.count).to eq(6)
-      expect(User.student.count).to eq(1)
     end
   end
 
   describe 'create from clever', :vcr do
     it 'gets data from clever' do
-      @district = District.create_from_clever('535e91b05fc8cb160e001645')
+      def create_district
+        @district = District.create_from_clever('535e91b05fc8cb160e001645')
+      end
+
+      expect { create_district }.not_to change { other_object_counts }
 
       expect(@district.name).to eq('#DEMO Clever Team Testing')
       expect(@district.valid?).to be_truthy
       expect(@district.id).to_not be_nil
-
-      expect(Classroom.count).to eq(4)
-      expect(User.teacher.count).to eq(6)
-      expect(User.student.count).to eq(1)
     end
   end