--- a/spec/models/lesson_planner_spec.rb
+++ b/spec/models/lesson_planner_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe LessonPlanner, :type => :model do