--- a/spec/models/school_spec.rb
+++ b/spec/models/school_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe School, type: :model do
   describe 'validations' do