--- a/spec/models/section_spec.rb
+++ b/spec/models/section_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Section, type: :model do