--- a/spec/models/student_profile_cache_spec.rb
+++ b/spec/models/student_profile_cache_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 
 describe StudentProfileCache, :type => :model do