--- a/spec/models/topic_specs.rb
+++ b/spec/models/topic_specs.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Topic, :type => :model do