--- a/spec/models/workbook_specs.rb
+++ b/spec/models/workbook_specs.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Workbook, :type => :model do