--- a/spec/serializers/activity_serializer_spec.rb
+++ b/spec/serializers/activity_serializer_spec.rb
@@ -1,33 +1,42 @@
 require 'spec_helper'
 
 describe ActivitySerializer, type: :serializer do
+  let(:activity)   { FactoryGirl.create(:activity) }
+  let(:serializer) { ActivitySerializer.new(activity) }
 
-  let!(:activity) { FactoryGirl.create(:activity) }
-  let!(:serializer) { ActivitySerializer.new(activity) }
+  describe '#to_json output' do
+    let(:json)   { serializer.to_json }
+    let(:parsed) { JSON.parse(json) }
 
+    activity_key = 'activity'
 
-  context "has expected attributes" do
-
-    let!(:parsed) { JSON.parse(serializer.to_json) }
-
-    it "should contain a root attribute" do
-      expect(parsed.keys).to include('activity')
+    it "includes '#{activity_key}' key" do
+      expect(parsed.keys).to include(activity_key)
     end
 
-    context "activity object" do
-      let!(:activity_json) { parsed['activity'] }
+    describe "'#{activity_key}' object" do
+      let(:parsed_activity) { parsed[activity_key] }
 
-      it "should have these keys" do
-        expected = ["uid", "id", "name", "description", "flags", "data", "created_at", "updated_at", "anonymous_path", "classification", "topic"]
-        expect(activity_json.keys).to match_array(expected)
+      topic_key = 'topic'
+
+      it "has the correct keys" do
+        expect(parsed_activity.keys)
+          .to match_array %w(anonymous_path
+                             classification
+                             created_at
+                             data
+                             description
+                             flags
+                             id
+                             name) +
+                            [topic_key] +
+                          %w(uid
+                             updated_at)
       end
 
-      it "should include a topic" do
-        expect(activity_json['topic'].class).to eq(Hash)
+      it "includes '#{topic_key}' Hash" do
+        expect(parsed_activity[topic_key]).to be_a(Hash)
       end
-
     end
-
   end
-
 end