--- a/spec/serializers/workbook_serializer_spec.rb
+++ b/spec/serializers/workbook_serializer_spec.rb
@@ -1,28 +1,28 @@
 require 'spec_helper'
 
 describe WorkbookSerializer, type: :serializer do
-  let!(:workbook) { FactoryGirl.create(:workbook) }
-  let!(:serializer) { WorkbookSerializer.new(workbook) }
+  let(:workbook)   { FactoryGirl.create(:workbook) }
+  let(:serializer) { WorkbookSerializer.new(workbook) }
 
+  describe '#to_json output' do
+    let(:json)   { serializer.to_json }
+    let(:parsed) { JSON.parse(json) }
 
-  context "has expected attributes" do
+    workbook_key = 'workbook'
 
-    let!(:parsed) { JSON.parse(serializer.to_json) }
-
-    it "should contain a root attribute" do
-      expect(parsed.keys).to include('workbook')
+    it "includes '#{workbook_key}' key" do
+      expect(parsed.keys).to include(workbook_key)
     end
 
-    context "workbook object" do
-      let!(:workbook_json) { parsed['workbook'] }
+    describe "'#{workbook_key}' object" do
+      let(:parsed_workbook) { parsed[workbook_key] }
 
-      it "should have these keys" do
-        expected = ["id", "title", "created_at", "updated_at"]
-        expect(workbook_json.keys).to match_array(expected)
+      it 'has the correct keys' do
+        expect(parsed_workbook.keys).to match_array %w(created_at
+                                                       id
+                                                       title
+                                                       updated_at)
       end
-
-
     end
-
   end
 end