--- a/spec/support/shared/student.rb
+++ b/spec/support/shared/student.rb
@@ -1,5 +1,3 @@
-require 'spec_helper'
-
 shared_examples_for "student" do