--- a/spec/support/shared/teacher.rb
+++ b/spec/support/shared/teacher.rb
@@ -1,5 +1,3 @@
-require 'spec_helper'
-
 shared_examples_for "teacher" do
 
   let(:teacher) { FactoryGirl.build(:classroom) }