--- a/spec/support/shared/uid.rb
+++ b/spec/support/shared/uid.rb
@@ -1,5 +1,3 @@
-require 'spec_helper'
-
 shared_examples_for "uid" do