in `spec/rails_helper.rb`

On my mid-2012 MacBook Pro, `bin/bundle exec rspec` went from

    Finished in 2 minutes 48.8 seconds (files took 22.56 seconds to load)
    262 examples, 0 failures, 4 pending
to

    Finished in 10.95 seconds (files took 27.63 seconds to load)
    262 examples, 0 failures, 4 pending

This branch is (re)based on(to) and, thus, includes PRs...
* [#522 - Empty db before running specs](https://github.com/empirical-org/Empirical-Core/pull/522)
* [#521 - Add spec/rails_helper.rb](https://github.com/empirical-org/Empirical-Core/pull/521)
* [#548 - Fix spurious serializer spec test failures](https://github.com/empirical-org/Empirical-Core/pull/548)